+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Some examples, more below in the actual changelog (newer entries are more likely
 
 ### Added
 * `WithRequestOptions` API option to configure default request options (#361, @anx-mschaefer)
+* kubernetes/v1: Add `autoscaling` attribute via `EnableAutoscaling` field (#369, @nachtjasmin)
 
 ### Changed
 * (internal) add "error-return" to request option interfaces (#361, @anx-mschaefer)